+++ rrset.go
@@ -1,10 +1,12 @@
 package udnssdk
 
 import (
-	"encoding/json"
 	"fmt"
 	"log"
 	"time"
+
+	"github.com/fatih/structs"
+	"github.com/mitchellh/mapstructure"
 )
 
 // RRSetsService provides access to RRSet resources
@@ -53,13 +55,24 @@ func (rp RawProfile) GetProfileObject() (interface{}, error) {
 	}
 }
 
-// remarshalJSON takes a structure and marshals it into another type
-func remarshalJSON(in, result interface{}) error {
-	bs, err := json.Marshal(in)
+// decode takes a RawProfile and uses reflection to convert it into the
+// given Go native structure. val must be a pointer to a struct.
+// This is identical to mapstructure.Decode, but uses the `json:` tag instead of `mapstructure:`
+func decodeProfile(m interface{}, rawVal interface{}) error {
+	config := &mapstructure.DecoderConfig{
+		Metadata:         nil,
+		TagName:          "json",
+		Result:           rawVal,
+		ErrorUnused:      true,
+		WeaklyTypedInput: true,
+	}
+
+	decoder, err := mapstructure.NewDecoder(config)
 	if err != nil {
 		return err
 	}
-	return json.Unmarshal(bs, &result)
+
+	return decoder.Decode(m)
 }
 
 // DirPoolProfile extracts the full Profile as a DirPoolProfile or returns an error
@@ -69,7 +82,7 @@ func (rp RawProfile) DirPoolProfile() (DirPoolProfile, error) {
 	if c != DirPoolSchema {
 		return result, fmt.Errorf("RDPoolProfile has incorrect JSON-LD @context %s\n", c)
 	}
-	err := remarshalJSON(rp, &result)
+	err := decodeProfile(rp, &result)
 	return result, err
 }
 
@@ -80,7 +93,7 @@ func (rp RawProfile) RDPoolProfile() (RDPoolProfile, error) {
 	if c != RDPoolSchema {
 		return result, fmt.Errorf("RDPoolProfile has incorrect JSON-LD @context %s\n", c)
 	}
-	err := remarshalJSON(rp, &result)
+	err := decodeProfile(rp, &result)
 	return result, err
 }
 
@@ -91,7 +104,7 @@ func (rp RawProfile) SBPoolProfile() (SBPoolProfile, error) {
 	if c != SBPoolSchema {
 		return result, fmt.Errorf("SBPoolProfile has incorrect JSON-LD @context %s\n", c)
 	}
-	err := remarshalJSON(rp, &result)
+	err := decodeProfile(rp, &result)
 	return result, err
 }
 
@@ -102,36 +115,35 @@ func (rp RawProfile) TCPoolProfile() (TCPoolProfile, error) {
 	if c != TCPoolSchema {
 		return result, fmt.Errorf("TCPoolProfile has incorrect JSON-LD @context %s\n", c)
 	}
-	err := remarshalJSON(rp, &result)
+	err := decodeProfile(rp, &result)
 	return result, err
 }
 
-// RawProfile converts to a naive RawProfile
-func (p DirPoolProfile) RawProfile() (RawProfile, error) {
-	var rp RawProfile
-	err := remarshalJSON(p, &rp)
-	return rp, err
+// encodeProfile takes a struct and converts to a RawProfile
+func encodeProfile(rawVal interface{}) RawProfile {
+	s := structs.New(rawVal)
+	s.TagName = "json"
+	return s.Map()
 }
 
 // RawProfile converts to a naive RawProfile
-func (p RDPoolProfile) RawProfile() (RawProfile, error) {
-	var rp RawProfile
-	err := remarshalJSON(p, &rp)
-	return rp, err
+func (p DirPoolProfile) RawProfile() RawProfile {
+	return encodeProfile(p)
 }
 
 // RawProfile converts to a naive RawProfile
-func (p SBPoolProfile) RawProfile() (RawProfile, error) {
-	var rp RawProfile
-	err := remarshalJSON(p, &rp)
-	return rp, err
+func (p RDPoolProfile) RawProfile() RawProfile {
+	return encodeProfile(p)
 }
 
 // RawProfile converts to a naive RawProfile
-func (p TCPoolProfile) RawProfile() (RawProfile, error) {
-	var rp RawProfile
-	err := remarshalJSON(p, &rp)
-	return rp, err
+func (p SBPoolProfile) RawProfile() RawProfile {
+	return encodeProfile(p)
+}
+
+// RawProfile converts to a naive RawProfile
+func (p TCPoolProfile) RawProfile() RawProfile {
+	return encodeProfile(p)
 }
 
 // DirPoolProfile wraps a Profile for a Directional Pool
@@ -145,7 +157,7 @@ type DirPoolProfile struct {
 
 // DPRDataInfo wraps the rdataInfo object of a DirPoolProfile response
 type DPRDataInfo struct {
-	AllNonConfigured bool    `json:"allNonConfigured,omitempty"`
+	AllNonConfigured bool     `json:"allNonConfigured,omitempty"`
 	IPInfo           *IPInfo  `json:"ipInfo,omitempty"`
 	GeoInfo          *GeoInfo `json:"geoInfo,omitempty"`
 }