+++ rrset_test.go
@@ -3,6 +3,7 @@ package udnssdk
 import (
 	"encoding/json"
 	"os"
+	"reflect"
 	"strings"
 	"testing"
 )
@@ -110,11 +111,7 @@ func Test_RRSets_Create(t *testing.T) {
 		Order:       "ROUND_ROBIN",
 		Description: "T. migratorius",
 	}
-	rp, err := p.RawProfile()
-	if err != nil {
-		t.Fatal(err)
-	}
-	val.Profile = rp
+	val.Profile = p.RawProfile()
 	t.Logf("Create(%v, %v)", r, val)
 	resp, err := testClient.RRSets.Create(r, val)
 
@@ -190,11 +187,7 @@ func Test_RRSets_Update(t *testing.T) {
 		Order:       "RANDOM",
 		Description: "T. migratorius",
 	}
-	rp, err := p.RawProfile()
-	if err != nil {
-		t.Fatal(err)
-	}
-	val.Profile = rp
+	val.Profile = p.RawProfile()
 	t.Logf("Update(%v, %v)", r, val)
 	resp, err := testClient.RRSets.Update(r, val)
 
@@ -318,4 +311,130 @@ func Test_RRSet_SelectPost(t *testing.T) {
 		t.Fatal(err)
 	}
 	t.Logf("RRSets: %+v\n", rrsets)
+}
+
+func Test_DirPoolProfile_RawProfile(t *testing.T) {
+	p := DirPoolProfile{}
+	expected := RawProfile{
+		"@context":    ProfileSchema(""),
+		"description": "",
+		"rdataInfo":   []DPRDataInfo(nil),
+	}
+
+	actual := p.RawProfile()
+
+	if !reflect.DeepEqual(actual, expected) {
+		for k := range expected {
+			if !reflect.DeepEqual(actual[k], expected[k]) {
+				t.Fatalf("RawProfile[%v] want %#v got %#v", k, expected[k], actual[k])
+			}
+		}
+		for k := range actual {
+			if !reflect.DeepEqual(actual[k], expected[k]) {
+				t.Fatalf("RawProfile[%v] want %#v got %#v", k, expected[k], actual[k])
+			}
+		}
+		t.Fatalf("want %#v got %#v", expected, actual)
+	}
+}
+
+func Test_RDPoolProfile_RawProfile(t *testing.T) {
+	p := RDPoolProfile{}
+	expected := RawProfile{
+		"@context":    ProfileSchema(""),
+		"description": "",
+		"order":       "",
+	}
+
+	actual := p.RawProfile()
+
+	if !reflect.DeepEqual(actual, expected) {
+		for k := range expected {
+			if !reflect.DeepEqual(actual[k], expected[k]) {
+				t.Fatalf("RawProfile[%v] want %#v got %#v", k, expected[k], actual[k])
+			}
+		}
+		for k := range actual {
+			if !reflect.DeepEqual(actual[k], expected[k]) {
+				t.Fatalf("RawProfile[%v] want %#v got %#v", k, expected[k], actual[k])
+			}
+		}
+		t.Fatalf("want %#v got %#v", expected, actual)
+	}
+}
+
+func Test_SBPoolProfile_RawProfile(t *testing.T) {
+	p := SBPoolProfile{}
+	expected := RawProfile{
+		"@context":      ProfileSchema(""),
+		"description":   "",
+		"actOnProbes":   false,
+		"rdataInfo":     []SBRDataInfo(nil),
+		"runProbes":     false,
+		"backupRecords": []BackupRecord(nil),
+	}
+
+	actual := p.RawProfile()
+
+	if !reflect.DeepEqual(actual, expected) {
+		for k := range expected {
+			if !reflect.DeepEqual(actual[k], expected[k]) {
+				t.Fatalf("RawProfile[%v] want %#v got %#v", k, expected[k], actual[k])
+			}
+		}
+		for k := range actual {
+			if !reflect.DeepEqual(actual[k], expected[k]) {
+				t.Fatalf("RawProfile[%v] want %#v got %#v", k, expected[k], actual[k])
+			}
+		}
+		t.Fatalf("want %#v got %#v", expected, actual)
+	}
+}
+
+func Test_RawProfile_TCPoolProfile(t *testing.T) {
+	input := RawProfile{
+		"@context":    TCPoolSchema,
+		"description": "",
+		"actOnProbes": false,
+		"rdataInfo":   []SBRDataInfo{},
+		"runProbes":   false,
+	}
+	expected := TCPoolProfile{
+		Context:     TCPoolSchema,
+		Description: "",
+		ActOnProbes: false,
+		RunProbes:   false,
+		RDataInfo:   []SBRDataInfo{},
+	}
+
+	actual, err := input.TCPoolProfile()
+
+	if err != nil {
+		t.Fatal(err)
+	}
+	if !reflect.DeepEqual(actual, expected) {
+		t.Fatalf("want %#v got %#v", expected, actual)
+	}
+}
+
+func Test_TCPoolProfile_RawProfile(t *testing.T) {
+	input := TCPoolProfile{}
+	expected := RawProfile{
+		"@context":    ProfileSchema(""),
+		"description": "",
+		"actOnProbes": false,
+		"rdataInfo":   []SBRDataInfo(nil),
+		"runProbes":   false,
+	}
+
+	actual := input.RawProfile()
+
+	if !reflect.DeepEqual(actual, expected) {
+		for k := range expected {
+			if !reflect.DeepEqual(actual[k], expected[k]) {
+				t.Fatalf("RawProfile[%v] want %#v got %#v", k, expected[k], actual[k])
+			}
+		}
+		t.Fatalf("want %#v got %#v", expected, actual)
+	}
 }